Modified project structure & implemented core gameplay
Now files are seperated into folders to clarify project structure
Fixed board generator
Implemented core gameplay mechanic that checks word presence and gives points for giving present words
Added simple points viewer in gameplay view
Added fundamentals for options in UI
Player points are being saved now
