+++ src/board.cpp
@@ -75,6 +75,8 @@ Board::Board(QObject *parent, std::vector<QString> wordsList, QVector<QChar> let
     }
     minY = maxY = y;
 
+    presentWords.push_back(lWord);
+
     tips.push_back({x, y}); //wstaw kratkę po właściwym słowie
     usedWordsCount++;
 
@@ -96,9 +98,10 @@ Board::Board(QObject *parent, std::vector<QString> wordsList, QVector<QChar> let
         qDebug() << "Chosen crossLetter:  " << crossLetter;
 
         if(unsuccessfulTries >= maxTries){
-                qDebug() << "throwing away word: " << lWord;
-                unsuccessfulTries = 0;
-                continue;
+            removedWords.push_back(lWord);
+            qDebug() << "throwing away word: " << lWord;
+            unsuccessfulTries = 0;
+            continue;
         }
 
         if(signMap[crossLetter].size() == 0){
@@ -142,6 +145,11 @@ Board::Board(QObject *parent, std::vector<QString> wordsList, QVector<QChar> let
                 }
             }
 
+            //check if crossLetter has been already used
+            if(std::find(tips.begin(), tips.end(), (std::pair<int, int>){x,y}) != tips.end()){
+                ok = false;
+            }
+
             //jak nie można dać słowa bo koliduje to wróć słowo do words
             if(!ok) {
 
@@ -157,12 +165,14 @@ Board::Board(QObject *parent, std::vector<QString> wordsList, QVector<QChar> let
                 }
                 else{
                     //poddaj się i wywal słowo
+                    removedWords.push_back(lWord);
                     qDebug() << "throwing away word: " << lWord;
                     unsuccessfulTries = 0;
                     continue;
                 }
             }
 
+            presentWords.push_back(lWord);
             tips.push_back({x,y});
             usedWordsCount++;
 
@@ -222,6 +232,11 @@ Board::Board(QObject *parent, std::vector<QString> wordsList, QVector<QChar> let
                 }
             }
 
+            //check if crossLetter has been already used
+            if(std::find(tips.begin(), tips.end(), (std::pair<int, int>){x,y}) != tips.end()){
+                ok = false;
+            }
+
             if(!ok) {
                 if(unsuccessfulTries < maxTries){
                     words.push_back(lWord);
@@ -236,15 +251,18 @@ Board::Board(QObject *parent, std::vector<QString> wordsList, QVector<QChar> let
                 }
                 else{
                     unsuccessfulTries = 0;
+                    removedWords.push_back(lWord);
                     //poddaj się i wywal słowo
                     qDebug() << "throwing away word: " << lWord;
                     continue;
                 }
             }
 
+            presentWords.push_back(lWord);
             usedWordsCount++;
             tips.push_back({x,y});
             unsuccessfulTries = 0;
+
             for(int i = 1; i <= first.length(); i++){
 
                 int xx = x - i;
@@ -305,14 +323,48 @@ int Board::Height() const
     return maxY - minY + 1;
 }
 
+std::vector<QString> &Board::PresentWords()
+{
+    return presentWords;
+}
+
+std::vector<QString> &Board::RemovedWords()
+{
+    return removedWords;
+}
+
+std::unordered_set<QString> &Board::AlreadyGuessedWords()
+{
+    return alreadyGuessedWords;
+}
+
+void Board::addGuessedWord(QString &guessedWord)
+{
+    alreadyGuessedWords.insert(guessedWord);
+}
+
 QVector<QChar>& Board::Letters()
 {
     return letters;
 }
 
-std::unordered_map<QString, std::vector<std::pair<int, int> > > &Board::WordMap() const
+std::unordered_map<QString, std::vector<std::pair<int, int>>>& Board::WordMap()
 {
     return wordMap;
+}
+
+//return 0 if words is not present
+//return 1 if word is present on a board
+//return 2 if word like that exists but it is not on the board
+int Board::checkWord(QString &wordToCheck)
+{
+    if(std::find(presentWords.begin(), presentWords.end(), wordToCheck) != presentWords.end()){
+        return 1;
+    }
+    if(std::find(removedWords.begin(), removedWords.end(), wordToCheck) != removedWords.end()){
+        return 2;
+    }
+    return 0;
 }
 
 void Board::operator=(const Board &obj)
@@ -324,6 +376,8 @@ void Board::operator=(const Board &obj)
         maxY = obj.maxY;
         usedWordsCount = obj.usedWordsCount;
         words = obj.words;
+        presentWords = obj.presentWords;
+        removedWords = obj.removedWords;
         signMap = obj.signMap;
         wordMap = obj.wordMap;
         letters = obj.letters;
