+++ src/board.h
@@ -4,6 +4,7 @@
 #include <QObject>
 #include <vector>
 #include <QString>
+#include <unordered_set>
 
 using Scheme = std::vector<std::vector<QChar>>;
 
@@ -17,7 +18,8 @@ private:
     int usedWordsCount = 0;
     //std::vector<QString> words = {"pobić", "opić", "bić", "obić", "poić", "pić"}, constWords;
     //std::vector<QString> words = {"kwit", "kotwica", "kwiat", "taki", "wacik", "koci", "kita", "kwita", "ciotka", "kwota", "owak", "owca"}, constWords;
-    std::vector<QString> words, allWords;
+    std::vector<QString> words, presentWords, removedWords;
+    std::unordered_set<QString> alreadyGuessedWords;
     std::unordered_map<QString, std::vector<std::pair<int, int>>> wordMap;
     std::unordered_map<QChar, std::vector<std::pair<int, int>>> signMap;
     QVector<QChar> letters;
@@ -33,9 +35,13 @@ public:
     int UsedWordsCount() const;
     int Width() const;
     int Height() const;
+    std::vector<QString>& PresentWords();
+    std::vector<QString>& RemovedWords();
+    std::unordered_set<QString>& AlreadyGuessedWords();
+    void addGuessedWord(QString& guessedWord);
     QVector<QChar>& Letters();
-    std::unordered_map<QString, std::vector<std::pair<int, int>>>& WordMap() const;
-    std::vector<pair<int, int>> checkWord(QString wordToCheck);
+    std::unordered_map<QString, std::vector<std::pair<int, int>>>& WordMap();
+    int checkWord(QString& wordToCheck);
     void operator=(const Board& obj);
 
 signals:
