+++ src/mainwindow.h
@@ -3,7 +3,9 @@
 
 #include <QMainWindow>
 #include "board.h"
+#include "qgridlayout.h"
 #include "qpushbutton.h"
+#include "user.h"
 
 QT_BEGIN_NAMESPACE
 namespace Ui {
@@ -24,8 +26,14 @@ private:
     int width = 1920, height = 1080;
     QString projectPath;
     Board board;
+    user player;
+    double gridCellSize = 0;
+    QGridLayout* gridLayout = nullptr;
     std::vector<QPushButton*> letButtons;
+    QString userDataPath;
     void prepareGrid();
     void prepareLetterButtons();
+    void setupButtons();
+    void updateGrid(QString &newWord);
 };
 #endif // MAINWINDOW_H
