+++ user_data/user_stats.txt
@@ -1,2 +1,2 @@
 Level=3
-Points=7
+Points=51
